add display in main
